上传部分漏掉的文件

+++ simpleDemo/src/main/resources/init.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS `user_0`;
 CREATE TABLE if not exists `user_0`
 (
-    `userId`        INT         NOT NULL AUTO_INCREMENT,
+    `userId`        BIGINT         NOT NULL AUTO_INCREMENT,
     `userName`      VARCHAR(45) NOT NULL,
     `userPassword`  VARCHAR(45) NOT NULL,
     `userPassword_` VARCHAR(45) NOT NULL,
@@ -11,7 +11,7 @@ CREATE TABLE if not exists `user_0`
 DROP TABLE IF EXISTS `user_1`;
 CREATE TABLE if not exists `user_1`
 (
-    `userId`        INT         NOT NULL AUTO_INCREMENT,
+    `userId`        BIGINT         NOT NULL AUTO_INCREMENT,
     `userName`      VARCHAR(45) NOT NULL,
     `userPassword`  VARCHAR(45) NOT NULL,
     `userPassword_` VARCHAR(45) NOT NULL,
